requirements.txt
Django==3.2.16
djangorestframework==3.14
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.13.0
requests==2.28.1
drf-yasg==1.21.4
drf-util==1.0.17
django-filter==22.1